+++ tests/system/sleep.py
@@ -62,7 +62,7 @@ class TestSleep(unittest.TestCase):
 	def test_enums(self):
 		"""Test if all enums from __init__.py are implemented"""
 		implemented_states = list(self._sleep.state_machine.states)
-		enum_states = [state.value for state in habapp_rules.system.SleepState]
+		enum_states = [state.value for state in habapp_rules.system.SleepState] + ["initial"]
 		self.assertEqual(len(enum_states), len(implemented_states))
 		self.assertTrue(all(state in enum_states for state in implemented_states))